refactor tests for pystiche.optim (#373)
* refactor tests for pystiche.optim

* delete old
